add core endpoints
