+++ crates/vrrb_rpc/src/rpc/server.rs
@@ -17,66 +17,10 @@ use vrrb_core::{
 };
 
 use super::api::{CreateTxnArgs, FullMempoolSnapshot};
-use crate::rpc::api::{FullStateSnapshot, RpcServer};
-
-pub struct RpcServerImpl {
-    node_type: NodeType,
-    vrrbdb_read_handle: VrrbDbReadHandle,
-    mempool_read_handle_factory: MempoolReadHandleFactory,
-    events_tx: UnboundedSender<DirectedEvent>,
-}
-
-#[async_trait]
-impl RpcServer for RpcServerImpl {
-    async fn get_full_state(&self) -> Result<FullStateSnapshot, Error> {
-        let values = self.vrrbdb_read_handle.state_store_values();
-
-        Ok(values)
-    }
-
-    async fn get_full_mempool(&self) -> Result<FullMempoolSnapshot, Error> {
-        let values = self.mempool_read_handle_factory.values();
-
-        Ok(values)
-    }
-
-    async fn get_node_type(&self) -> Result<NodeType, Error> {
-        Ok(self.node_type)
-    }
-
-    async fn create_txn(&self, args: CreateTxnArgs) -> Result<(), Error> {
-        let txn = vrrb_core::txn::Txn::new(NewTxnArgs {
-            sender_address: args.sender_address,
-            sender_public_key: args.sender_public_key,
-            receiver_address: args.receiver_address,
-            token: args.token,
-            amount: args.amount,
-            payload: args.payload,
-            signature: args.signature,
-            validators: None,
-            nonce: args.nonce,
-        });
-
-        let event = Event::NewTxnCreated(txn);
-
-        self.events_tx
-            .send((Topic::Transactions, event))
-            .map_err(|err| {
-                telemetry::error!("could not queue transaction to mempool: {err}");
-                Error::Custom(err.to_string())
-            })?;
-
-        Ok(())
-    }
-
-    async fn create_account(&self, account: Account) -> Result<(), Error> {
-        todo!()
-    }
-
-    async fn update_account(&self, account: Account) -> Result<(), Error> {
-        todo!()
-    }
-}
+use crate::rpc::{
+    api::{FullStateSnapshot, RpcServer},
+    server_impl::RpcServerImpl,
+};
 
 #[derive(Debug, Clone)]
 pub struct JsonRpcServerConfig {
